requirements.txt
py-cord[voice]
google-generativeai
google-cloud-texttospeech
discord